ref: code moved to helper.js and jsdoc added
