+++ src/reducers/index.js
@@ -1,2 +1,3 @@
 export { productsReducer, sortByReducer } from "./products-reducer";
 export { loginReducer, registerReducer } from "./auth-reducer";
+export { wishlistReducer } from "./wishlist-reducer";
