+++ src/utils/helpers.js
@@ -1,2 +1,19 @@
+/**
+ * Removed element from array
+ * @param {Array} arr Array from which we need to remove an element
+ * @param {any} element Element that needs to be removed from arr
+ * @returns Array with element removed
+ */
 const removeFromArray = (arr, element) =>
 	arr.filter((item) => item !== element);
+
+/**
+ * Check if its present in the array
+ * @param {Array} arr
+ * @param {any} element Element that needs to be searched from arr
+ * @returns true if element is found else false
+ */
+const presentInArray = (arr, element) =>
+	arr.find((item) => item === element) !== undefined ? true : false;
+
+export { removeFromArray, presentInArray };
